pyproject.toml
[tool.poetry]
name = "newsletter"
version = "0.1.0"
description = ""
authors = ["Baha Dzhalalov <bahadzhalalov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.21.0"
pymysql = "^1.0.2"
drf-yasg = "^1.21.4"


[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.core.masonry.api"
